requirements.txt
# Core LLM APIs
anthropic==0.25.5
openai==1.30.1         # For GPT-3.5 / GPT-4o (future use)
google-generativeai==0.4.1  # For Gemini (future use)

# LangGraph for agent orchestration
langgraph==0.0.18

# Retrieval support
requests==2.31.0

# Env management
python-dotenv==1.0.1

# Optional utils
httpx==0.27.0           # Fix for Claude internal HTTP issue
fastapi
uvicorn
sqlalchemy
alembic
python-dateutil
sentence-transformers>=2.2.2
